### 2025-04-02 #7984 @editorialbot
[PRE REVIEW]: TransiFlow: A Python package for performing bifurcation analysis on fluid flow problems
**Submitting author:** @Sbte (<a href="http://orcid.org/0000-0002-4228-6479">Sven Baars</a>)
**Repository:** https://github.com/BIMAU/transiflow
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
