### 2025-06-16 #7984 @jatkinson1000
Hi @YidanXue Thanks for that.
The JOSS review is checklist based with further comments via GitHub.
Whilst you are welcome to look at the code this week please note that we will need two more reviewers

You can read a bit more about the process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and feel free to ask me any questions.
