### 2025-06-16 #7984 @jodemaey
Hi @jatkinson1000 ,

I agree on reviewing this paper.
However, please note that I have myself currently a paper under review for a bifurcation related code (working on top of the AUTO software), see https://github.com/openjournals/joss-reviews/issues/8079, so I don't know if it constitute a conflict of interest.
I also co-authored a paper with Henk Dijkstra in 2024.

What do you think ?

Cheers,

Jonathan
