### 2025-06-23 #7984 @jatkinson1000
Hi there @ali-ramadhan, @luohancfd, and @cwentland0!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a finite volume CFD code written in Python with implementation of several test cases allowing users to investigate bifurcation analysis in flows.

Would any of you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
