### 2025-06-29 BIMAU/transiflow#57 @YidanXue
[JOSS review] Some minor comments
These comments relate to my review: https://github.com/openjournals/joss-reviews/issues/8453

The paper is very nicely written, as I have checked off most items in my review checklist. Here are just some minor comments for the authors to consider.

1. The ‘Installation’ section in README.md should include the step of cloning the repository, and clarify that the user needs to be inside the ‘transiflow’ source directory before running ‘pip install -e .’.
2. The authors should provide more detailed instructions on how to reproduce Figure 1. The qg.py gave me similar results, but not the same.
3. It would be helpful if the example problems in the ‘examples’ folder included a brief explanation of the expected outputs. For instance, after running ‘amoc.py’, I got several .npy and .params files without any figure or guidance on how to interpret them, which was confusing (for someone like me from a different field).
