### 2025-07-14 #8453 @jatkinson1000
Hi folks, given @Sbte's availability I'll set a reminder for 4th August to see how things are going.
Please to progress with your checklists, opening issues on the repository where possible and including a link to this page so I can see what's being discussed.

Let me know if you have an issues during this time, thanks!
