### 2025-07-16 BIMAU/fvm#58 @svchb
[JOSS-Review] Comments on the paper
Following comments from my side regarding the paper https://github.com/openjournals/joss-reviews/issues/8453:

1) There is insufficient explanation of what bifurcation (analysis) is.
2) The types of bifurcation used in the figure caption are not mentioned anywhere else.
3) What about BifurcationKit.jl which seems to be actively developed?

Mistakes:
- “such as **a** lid‑driven and differentially heated cavities-> “such as lid‑driven and differentially heated cavities
- “Examples are dynamical systems from ocean models…” -> “Examples include dynamical systems from ocean models…”
- The last sentence of the figure caption are actually two sentences
- Titles in your reference list are inconsistently capitalized
