### 2025-07-25 BIMAU/transiflow#59 @cwentland0
[JOSS-Review] Review comments
All around a good paper and very well-written code. A couple comments and questions to address for review openjournals/joss-reviews#8453:
- [ ] I agree with @svchb that there is insufficient explanation of the utility/importance of bifurcation analysis.
- [ ] What were the roles of the co-authors (David, Fred, and Henk) in writing this code/paper, or relation to the first author? Not clear from the commit history.
- [ ] The paper should make it clear that the code, as it is currently written, only handles solutions for the incompressible Navier-Stokes equations with optional temperature/salinity transport. References to general "fluid flow" are too broad.
- [ ] Either the README or the documentation should provide guidelines for community contributions, report issues, etc.
- [ ] The installation instructions or documentation should include a guide for installing optional dependencies to access the `Epetra`, `HYMLS`, and `PETSc` backends. I have no way of testing their validity beyond backing out the installation process from the GitHub actions.
