### 2025-08-11 d0f70e8
paper: Process reviewer comments.
