### 2025-08-11 4aebb8d
paper: Process reviewer comments.
