+++ paper/paper.bib
@@ -30,7 +30,7 @@
   author =       {A. Dhooge and W. Govaerts and Yu. A. Kuznetsov and
                   H. G.E. Meijer and B. Sautois},
   title =        {New Features of the Software
-                  \sc{MatCont} for Bifurcation
+                  {M}at{C}ont for Bifurcation
                   Analysis of Dynamical Systems},
   journal =      {Mathematical and Computer Modelling of Dynamical
                   Systems},
@@ -99,6 +99,22 @@
   issn =         {1793-6551},
   month =        may,
   publisher =    {World Scientific Pub Co Pte Lt},
+}
+
+@Article{sapsis:13,
+  author =       {Themistoklis P. Sapsis and Henk A. Dijkstra},
+  title =        {Interaction of Additive Noise and Nonlinear Dynamics
+                  in the Double-Gyre Wind-Driven Ocean Circulation},
+  journal =      {Journal of Physical Oceanography},
+  volume =       43,
+  number =       2,
+  pages =        {366-381},
+  year =         2013,
+  doi =          {10.1175/jpo-d-12-047.1},
+  url =          {http://dx.doi.org/10.1175/jpo-d-12-047.1},
+  issn =         {1520-0485},
+  month =        feb,
+  publisher =    {American Meteorological Society},
 }
 
 @phdthesis{song:19,
