### 2025-08-11 9ed4005
README: Add more detailed installation instructions.
For #57.
