### 2025-08-11 8a8adae
README: Add more detailed installation instructions.
For #57.
