### 2025-08-11 0c859ca
paper: Add BifurcationKit.jl.
