+++ paper/paper.bib
@@ -169,6 +169,24 @@
   publisher =    {Global Science Press},
 }
 
+@misc{veltz:20,
+  author =       {Veltz, Romain},
+  eprint =       {hal-02902346},
+  eprinttype =   {hal},
+  hal_id =       {hal-02902346},
+  hal_version =  {v1},
+  institution =  {{Inria Sophia-Antipolis}},
+  keywords =     {pseudo-arclength-continuation ; periodic-orbits ;
+                  floquet ; gpu ; bifurcation-diagram ; deflation ;
+                  newton-krylov},
+  month =        Jul,
+  pdf =
+                  {https://hal.archives-ouvertes.fr/hal-02902346/file/354c9fb0d148262405609eed2cb7927818706f1f.tar.gz},
+  title =        {{BifurcationKit.jl}},
+  url =          {https://hal.archives-ouvertes.fr/hal-02902346},
+  year =         2020,
+}
+
 @Article{westen:24,
   author =       {Ren{\'e} M. van Westen and Michael Kliphuis and Henk
                   A. Dijkstra},
