### 2025-08-11 1d807a5
paper: Add BifurcationKit.jl.
