### 2025-08-11 BIMAU/transiflow#59 @Sbte
>  What were the roles of the co-authors (David, Fred, and Henk) in writing this code/paper, or relation to the first author? Not clear from the commit history.

I didn't see any guidance on how to add this to the paper, but I think I had to add it during the submission process. But for completeness sake, I'll add it here:

David contributed the PETSc backend, as well as various other useful bits of the software. Fred and Henk wrote both the Fortran code and the book that this package is based on, as well as providing guidance, verification, and endless discussions to make sure the code is correct.
