### 2025-08-11 104ddfd
Add contribution guidelines.
For #59.
