### 2025-08-11 4dd9455
Add contribution guidelines.
For #59.
