### 2025-08-11 c93dfdd
README: Add instructions for installing optional backends.
For #59.
