### 2025-08-11 e6ca98e
README: Add instructions for installing optional backends.
For #59.
