### 2025-08-30 13b578f
paper: Process Fred's comments.
For #58, #59.
