### 2025-08-30 57a2875
paper: Process Fred's comments.
For #58, #59.
