### 2025-08-30 e25bb4c
examples: Move docstrings to module level.
For #57.
