### 2025-08-30 1be400e
examples: Move docstrings to module level.
For #57.
