### 2025-08-30 8c08600
docs: Add a note one the purpose of the examples.
For #57.
