### 2025-08-30 f380bb3
docs: Add a note one the purpose of the examples.
For #57.
