### 2025-09-04 78379de
docs: Add a section on the examples.
