### 2025-09-04 cbfcf17
docs: Add a section on the examples.
