+++ docs/index.rst
@@ -16,6 +16,7 @@ The package is fully agnostic of the computational backend, which allows for eas
 
    continuation
    backends
+   examples
    custom-bc
    custom-model
 
