### 2025-09-04 BIMAU/transiflow#58 @Sbte
The issues described here should have been addressed by #60 and #61. Feel free to reopen this issue if the resolution was inadequate.

Thanks especially for the BifurcationKit.jl recommendation (this was unfortunately not available yet when we started development of this software, and serves a slightly different purpose, but may have been nice to use if it were around back then), and also thanks for the last comment. This seems to be a bug in the software used to compile the paper. We worked around the issue in the new version.
