### 2025-09-04 BIMAU/transiflow#57 @Sbte
The issues described here should have been addressed by #60 and #61. Feel free to reopen this issue if the resolution was inadequate.
