### 2025-09-04 #8453 @jatkinson1000
Thanks for the update @Sbte

@svchb @YidanXue @cwentland0 please could you examine the changes that have been made to the code and see if you feel the software now satisfies the criteria for JOSS and continue your checklists.

If it is still falling short in any areas please raise these as issues, linking back to this thread, or post broader comments here.
If you could try and do this over the next week that would be great, thank you.
