### 2025-09-07 #8453 @YidanXue
@jatkinson1000 I have ticked all items in my checklist, after checking the GitHub repository. I was wondering if there is also an updated manuscript for us to look at?
