### 2025-09-07 #8453 @YidanXue
Thanks @jatkinson1000 ! I noticed some minor changes in the manuscript, some of which has addressed my comment on reproducing Figure 1. I am happy for this paper to be accepted for publication. Just two very minor comments if the authors would address before the publication:

1. The Sapsis, 2013 reference in the caption of Figure 1 doesn't display correctly on the pdf.
2. I would recommend the authors to separate their first paragraph (which is very long) into two paragraphs.
