### 2025-09-09 0f3a53e
paper: Minor suggestions.
