### 2025-09-16 2d9601b
Add JOSS editorial changes to improve readability.
