+++ paper/paper.md
@@ -30,30 +30,32 @@ bibliography: paper.bib
 
 # Summary
 
-Dynamical systems such as those derived from models of fluid flows show transition behavior associated with fluid flow instabilities [@wubs:23].
-Examples include dynamical systems from ocean models in which transition behavior is caused by slow changes in parameters representing the surface forcing [@westen:24].
-By flow transition one means a qualitative change in the flow when a specific parameter is changed, e.g. a transition from a no-flow heat-conducting fluid to a heat-transporting flow (as in Rayleigh-Bénard convection), or a flow which has a steady forcing and steady boundary conditions that turns from a steady flow into a transient flow which may even produce sound (as in the von Kármán vortex street).
-In the foregoing two examples both qualitatively different solutions do not just appear out of the blue.
-If one perturbs the steady solution before the transition point then with time the perturbation will die out and we will recover the steady solution.
-This means that the steady solution is stable.
-However, the perturbation already reveals the shape of the flow that will occur after the transition point.
-After this point that shape will grow into a steady flow (the no-flow solution becomes unstable) for the first example and a transient flow in the second example (the stationary flow becomes unstable).The parameter value for which the transition sets in is called a bifurcation point, which, in certain cases, is also referred to as tipping point.
+Dynamical systems derived from models of fluid flow show transition behaviors associated with fluid flow instabilities [@wubs:23].
+For example, dynamical systems from ocean models in which transition behavior is caused by slow changes in parameters representing the surface forcing [@westen:24].
+A flow transition is a qualitative change in the flow when a specific parameter is varied, e.g. a transition from a no-flow heat-conducting fluid to a heat-transporting flow (as in Rayleigh-Bénard convection), or a steady flow which has a steady forcing and steady boundary conditions that turns into a transient flow which may even produce sound (as in the von Kármán vortex street).
+In both these examples the qualitatively different solutions do not appear out of the blue.
 
-Studying these phenomena (bifurcation analysis) can be done by performing numerical simulations with a model and observing its transient behavior after a certain time.
-This is, however, computationally very expensive, and in many cases infeasible.
-Instead, so called continuation methods are able to trace stable and unstable steady states in parameter space, obviating expensive transient simulations [@dijkstra:05].
-The `TransiFlow` Python package implements a continuation framework in which fluid flow problems can be studied with the help of several computational back-ends that can, based on the needs of the user, easily be switched between.
+If one perturbs a steady solution before the transition point then the perturbation will die out and recover to the steady solution.
+This means that the steady solution is stable.
+However, this perturbed flow already reveals the shape that will occur after the transition point.
+For forcing beyond this point that shape will grow into a new steady flow (the stable solution becomes unstable) for Rayleigh-Bénard convection and into a transient flow for the von Kármán vortex street (the stationary flow becomes unstable).
+The parameter value for which the transition occurs is called a bifurcation point, sometimes also referred to as a transition or tipping point.
+
+Studying these phenomena (bifurcation analysis) can be done by performing numerical simulations and observing transient behavior after a certain time.
+This is, however, computationally expensive and in many cases infeasible.
+Instead, so-called continuation methods are able to trace stable and unstable steady states in parameter space, obviating expensive transient simulations [@dijkstra:05].
+The `TransiFlow` Python package implements a continuation framework in which fluid flow problems can be studied with the help of several computational back-ends that can, based on the needs of the user, be easily switched between.
 
 One motivation behind `TransiFlow` is that writing research software that works efficiently on a parallel computer is a challenging task.
[…]
