### 2025-10-05 fb4784d
paper: Add notes on how TransiFlow is being used for ongoing research.
