### 2025-10-05 e6a6c88
paper: Remove the 1% claim about the Python overhead.
