### 2026-01-16 #8453 @cwentland0
After reading the thread at BIMAU/transiflow#62, and a careful re-read of the paper after, I have a few points in response:
1. I find the mention of climate modeling fairly innocuous. It does not make a statement that this code is compatible with an Earth system model out-of-the-box, but is simply a motivating example of a large-scale problem for which a continuation method might be useful. Some simplified climate models are implemented in the examples, so I feel the link is useful.
2. The paper does not seem to make any generalized claims about being able to integrate arbitrary _models_, even though [looking at an example of another discretization](https://github.com/BIMAU/transiflow/blob/ec5d87de4d8aae7162c96ec2990e7668c095d000/transiflow/OceanDiscretization.py) does imply this is a pretty excruciating process. This seems to have just been a confusing diversion in BIMAU/transiflow#62.
3. On second consideration, the unqualified claim that models can be hooked up to the various backends is misleading, as the provided parallel backends do assume a Cartesian mesh. This is certainly _not_ true on typical unstructured Earth system model grids. I suggest that some sort of qualifying statement be made that the current implementation only permits problems that are defined on a Cartesian mesh.
