### 2026-01-17 #8453 @Sbte
> 2. The paper does not seem to make any generalized claims about being able to integrate arbitrary _models_, even though [looking at an example of another discretization](https://github.com/BIMAU/transiflow/blob/ec5d87de4d8aae7162c96ec2990e7668c095d000/transiflow/OceanDiscretization.py) does imply this is a pretty excruciating process. This seems to have just been a confusing diversion in [Joss paper editorial changes BIMAU/transiflow#62](https://github.com/BIMAU/transiflow/pull/62).

The OceanDiscretization you mention is the in-progress implementation of a full 3D ocean model, which of course is a bit complicated. Using it with an existing model should be less difficult.

> 3. On second consideration, the unqualified claim that models can be hooked up to the various backends is misleading, as the provided parallel backends do assume a Cartesian mesh. This is certainly _not_ true on typical unstructured Earth system model grids. I suggest that some sort of qualifying statement be made that the current implementation only permits problems that are defined on a Cartesian mesh.

This is true. However, I do not know if we make this claim anywhere in the paper or the documentation. I think we either mention modifying or extending the Discretization class, or supplying your own backend. In the first case, you'll have the Cartesian mesh, in the second case the mesh can be arbitrary. If we mention anywhere that you can use any model with the provided backends, that should be corrected.
