### 2026-01-23 #8453 @cwentland0
I am referring to the portion that reads as follows (emphasis mine):
> By abstracting away the computational back-end from the user, **the user can adjust a model to their own needs** on their own machine (e.g., a laptop) in Python using the SciPy back-end, and once the model works, **run a large scale simulation** on a supercomputer, e.g., using the Trilinos back-end, which can use a combination of OpenMP, MPI and potentially GPUs, **without requiring any changes to the code**.

To me, this certainly implies that the user can hook up an arbitrary model to the parallel backend. The failure to mention that the parallel backends assume a Cartesian mesh thus grossly overstates the generality of the code. There must be a sentence somewhere that makes this restriction 100% clear.
