### 2026-01-24 #8453 @Sbte
Ah, that makes sense, yeah. When writing that I was assuming adjusting one of the included models, which are all implemented on a Cartesian grid. Making that restriction more explicit sounds like a good idea. What about this:

> By abstracting away the computational back-end the user can adjust a model to their own needs on their own machine (e.g., a laptop) in Python using the SciPy back-end, and once the model works, and assuming it is implemented on a Cartesian grid, the user can then run a large scale simulation on a supercomputer using, e.g., the Trilinos back-end, which can use a combination of OpenMP, MPI, and potentially GPUs, without requiring any changes to the code.
