### 2026-02-06 #8453 @YidanXue
The sentence reads good, but quite long. So I would suggest to separate it into two sentences (sentence 1: By abstracting away...SciPy back-end. sentence 2: Once the model works...to the code.).
