### 2026-02-06 #8453 @svchb
I would propose the following:
```
By abstracting the computational back end, users can develop and test a model locally in Python (e.g., with a SciPy back end)
and then run large-scale simulations on HPC systems (e.g., with a Trilinos back end) without changing the model code,
provided the model is implemented on a Cartesian grid.
```
