### 2026-02-15 #8453 @Sbte
I tried to combine the two suggestions and committed the result. I now reads

> We achieve this by abstracting away the computational back-end, which allows the user to adjust a model to their own needs on their own machine (e.g., a laptop) in Python using the SciPy back-end. Then, once the model works, and assuming it is implemented on a Cartesian grid, the user can run a large-scale simulation on an HPC system using, e.g., the Trilinos back-end, which can use a combination of OpenMP, MPI, and potentially GPUs, without requiring any changes to the model code.
