### 2026-02-15 #8453 @svchb
I think you can leave this out:
> Then, once the model works, and
