### 2026-02-16 #8453 @jatkinson1000
Thanks all, please take a quick read of the above paper, regenerated with @Sbte's latest changes.
If you are happy with how it now reads and the statements it makes of the software's capabilities please drop me a 👍
