### 2026-03-01 #8453 @jatkinson1000
Super, now that there is consensus we can continue with [the checklist above](https://github.com/openjournals/joss-reviews/issues/8453#issuecomment-3295805161).

Specifically @Sbte we are waiting for:

please could you let me know when you have done these and I will perform editorial tasks and check them off.
