### 2026-03-06 #8453 @jatkinson1000
@Sbte let me know if you need assistance creating a release and archive.

Zenodo has capabilities for automatically archiving GitHub releases - documented [here](https://help.zenodo.org/docs/github/archive-software/github-upload/).

If you prefer not to use Zenodo or Figshare your institute may have the capability to host archives and provide a DOI.
