### 2026-03-09 #8453 @Sbte
> Double check authors and affiliations (including ORCIDs)

They are correct.

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

I created a release just now.

> Archive the release on Zenodo/figshare/etc and post the DOI here.

10.5281/zenodo.18926085

The guide was very helpful, thanks.

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

I did this. It seems like Zenodo is partially down at the moment, so I had to repeat my edits many times. It seems like they got through.

> Make sure that the license listed for the archive is the same as the software license.

It is.
