### 2026-03-13 #8453 @jatkinson1000
Set version with @editorialbot set v1.0.1 as version
