### 2026-03-13 #8453 @jatkinson1000
Thanks @Sbte

At the moment the names (middle initials) and affiliations of the Zenodo archive do not match those on the paper. This is a requirement for JOSS publication. Please update them to be consistent.

Updating either the archive or the paper is fine, whichever you feel is 'correct', but they must be consistent.
