### 2026-03-14 #8453 @Sbte
I updated the names on Zenodo.

The affiliation was harder. Once the entry field started searching for existing affiliations I would then only be able to select the one already present in their system. I solved this by blocking the request that searches for existing names in their system, which allowed me to add a new affiliation. This sounds unintended, but it was apparently the only way to add the affiliation as described in the paper.
