### 2026-03-21 #8453 @jatkinson1000
Thanks @Sbte this all looks good to me!
I'm recommending acceptance and publication and this will now be passed up to the track editor for final checks and release.

Thanks also to our dedicated volunteer reviewers @YidanXue, @svchb, @cwentland0
As a reminder, you can [register as a JOSS reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.

Thanks for your submission and participation in the open review process.

@Sbte thanks for the note about Zenodo affiliations and having to block and manually specify. I will aim to raise this with the editorial team for discussion.
