### 2026-03-31 #8453 @kyleniemeyer
@Sbte sorry for the delay, I just have a few editorial corretions before accepting:
- at the end of the Figure 1 caption "can be found in (Sapsis & Dijkstra, 2013)", please change the citation command to `@author:2001` rather than `[@author:2001]`, such that the in-text citation is formatted as "can be found in Sapsis & Dijkstra (2013)"
- same correction needed in a few places in the "Past and ongoing research" section.
- Please update the Bernuzzi et al. article details, as it has now been published: https://www.sciencedirect.com/science/article/pii/S0167278924003415
